feat: add sensor dashboard
